requirements.txt
absl-py==0.10.0
amqp==1.4.9
anyjson==0.3.3
appdirs==1.4.4
appnope==0.1.0
asgiref==3.2.0
asn1crypto==0.24.0
astroid==2.2.5
astunparse==1.6.3
attrs==19.1.0
authy==2.2.6
autobahn==19.8.1
Automat==0.7.0
autopep8==1.4.4
Babel==2.8.0
backcall==0.2.0
beautifulsoup4==4.9.1
billiard==3.3.0.23
bs4==0.0.1
cached-property==1.5.1
cachetools==4.1.1
celery==3.1.18
certifi==2019.9.11
cffi==1.12.3
chardet==3.0.4
colorama==0.4.1
constantly==15.1.0
cryptography==2.7
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
Django==2.2.4
django-ads==1.1.1
django-allauth==0.40.0
django-appconf==1.0.1
django-bootstrap-datepicker-plus==3.0.5
django-braces==1.14.0
django-cereal==0.1.2
django-classy-tags==1.0.0
django-cors-headers==3.4.0
django-countries==5.5
django-crispy-forms==1.7.2
django-durationfield==0.5.1
django-google-maps==0.12.1
django-googlemap==0.5
django-js-reverse==0.9.1
django-money==0.7.0
django-phone-verify==2.0.0
django-phonenumber-field==4.0.0
django-sekizai==1.1.0
django-tinymce4-lite==1.7.5
djangorestframework==3.11.0
et-xmlfile==1.0.1
gast==0.3.3
google-auth==1.18.0
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
googleads==3.5.0
grpcio==1.33.1
h5py==2.10.0
httpie==2.2.0
httplib2==0.18.1
hyperlink==19.0.0
idna==2.8
imageio==2.5.0
imgaug==0.2.9
importlib-metadata==2.0.0
incremental==17.5.0
ipykernel==5.3.3
ipython==7.16.1
ipython-genutils==0.2.0
isodate==0.6.0
isort==4.3.21
jdcal==1.4.1
jedi==0.17.2
joblib==0.16.0
jsmin==2.2.2
jupyter-client==6.1.6
jupyter-core==4.6.3
Keras==2.4.3
Keras-Preprocessing==1.1.2
keras-segmentation==0.3.0
kiwisolver==1.3.1
kombu==3.0.37
lazy-object-proxy==1.4.1
leaflet==0.0.3
lxml==4.5.2
Markdown==3.3.3
matplotlib==3.3.3
mccabe==0.6.1
mysqlclient==2.0.1
networkx==2.5
nexmo==2.4.0
numpy==1.18.5
oauth2client==4.1.3
oauthlib==3.1.0
opencv-python==4.4.0.46
openpyxl==3.0.4
opt-einsum==3.3.0
pandas==1.1.0
parso==0.7.0
pep8==1.7.1
pexpect==4.8.0
phonenumbers==8.12.7
pickleshare==0.7.5
Pillow==6.1.0
pipenv==2020.6.2
prompt-toolkit==3.0.5
protobuf==3.13.0
psycopg2==2.8.3
ptyprocess==0.6.0
py-moneyed==0.5.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.5.0
pycparser==2.19
Pygments==2.6.1
PyHamcrest==1.9.0
PyJWT==1.7.1
pylint==2.3.1
pyparsing==2.4.7
PySocks==1.7.1
python-dateutil==2.8.1
python-dotenv==0.14.0
python3-openid==3.1.0
pytz==2018.5
PyWavelets==1.1.1
PyYAML==5.3.1
pyzmq==19.0.1
redis==3.5.3
reportlab==3.5.44
requests==2.22.0
requests-oauthlib==1.2.0
requests-toolbelt==0.9.1
rsa==4.6
scikit-image==0.17.2
scikit-learn==0.23.1
scipy==1.5.1
Shapely==1.7.1
six==1.12.0
sklearn==0.0
sklearn-pandas==2.0.2
soupsieve==2.0.1
sqlparse==0.2.4
stripe==2.37.1
suds-jurko==0.6
tensorboard==2.3.0
tensorboard-plugin-wit==1.7.0
tensorflow==2.3.1
tensorflow-estimator==2.3.0
termcolor==1.1.0
threadpoolctl==2.1.0
tifffile==2020.10.1
tornado==6.0.4
tqdm==4.51.0
traitlets==4.3.3
twilio==6.44.0
txaio==18.8.1
typed-ast==1.4.0
urllib3==1.25.3
virtualenv==16.6.1
virtualenv-clone==0.5.4
virtualenvwrapper-win==1.2.5
wcwidth==0.2.5
Werkzeug==1.0.1
wrapt==1.11.2
xmltodict==0.12.0
zeep==3.4.0
zipp==3.4.0
zope.interface==4.6.0
